Updates to IRTranslator
