+++ src/tests/Microsoft.PowerFx.Interpreter.Tests/FlagTests.cs
@@ -46,88 +46,5 @@ namespace Microsoft.PowerFx.Core.Tests
 
             Assert.True(resultWithoutFlag.IsSuccess);
         }
-
-        [Theory]
-        [InlineData("X")]
-        [InlineData("$\"Test {X}\"")]
-        [InlineData("!X")]
-        [InlineData("First(Table(X))")]
-        [InlineData("X.Field1")]
-        [InlineData("X + 1")]
-        [InlineData("X + Date(2022, 10, 27)")]
-        [InlineData("X * 1")]
-        [InlineData("X < 2")]
-        [InlineData("\"The\" in X")]
-        [InlineData("123 in X")]
-        [InlineData("123 in Table(X)", false)]
-        [InlineData("IsBlank(X)")]
-        [InlineData("IsError(X)", false)]
-        [InlineData("Text(X)")]
-        [InlineData("Value(X)")]
-        [InlineData("Boolean(X)")]
-        [InlineData("Index([1,2,3].Value, X)")]
-
-        // Ensures expression binds without any errors - but issues a warning for the deferred(unknown) type.
-        public void DeferredTypeTest_EnableDeferredType(string script, bool isEvalAllowed = true)
-        {
-            Preview.FeatureFlags.StringInterpolation = true;
-            var symbolTable = new SymbolTable();
-            symbolTable.AddVariable("X", FormulaType.Unknown);
-            TestDeferredTypeBindingWarning(script, Features.EnableDeferredType, symbolTable, isEvalAllowed);
-        }
-
-        [Theory]
-        [InlineData("$\"Test {X} {R}\"", "Invalid argument type (Record). Expecting a Text value instead.")]
-        [InlineData("Table(X, N)", "Cannot use a non-record value in this context")]
-        [InlineData("X.field + N.field", "Invalid use of '.'")]
-        [InlineData("Index([1,2,3], X).missing", "Name isn't valid. 'missing' isn't recognized")]
-
-        // Ensures expression issues an error if it exists, despite the deferred type.
-        public void DeferredTypeTest_EnableDeferredType_Negative(string script, string errorMessage)
-        {
-            Preview.FeatureFlags.StringInterpolation = true;
-            var symbolTable = new SymbolTable();
-            symbolTable.AddVariable("X", FormulaType.Unknown);
-            symbolTable.AddVariable("N", FormulaType.Number);
-            symbolTable.AddVariable("R", RecordType.Empty());
-            TestBindingError(script, Features.EnableDeferredType, errorMessage, symbolTable);
-        }
-
-        private void TestDeferredTypeBindingWarning(string script, Features features, SymbolTable symbolTable = null, bool isEvalAllowed = true)
-        {
-            var config = new PowerFxConfig(features)
-            {
-                SymbolTable = symbolTable
-            };
-
-            var engine = new RecalcEngine(config);
-            var result = engine.Check(script);
-
-            Assert.True(result.IsSuccess);
-            Assert.True(result.Errors.Count() > 0);
-            Assert.True(result.Errors.All(error => error.MessageKey.Equals(TexlStrings.WarnUnknownType.Key)));
-
-            if (isEvalAllowed)
-            {
-                var evaluator = CheckResultExtensions.GetEvaluator(result);
-                var evaluatorResult = evaluator.Eval();
-                Assert.IsType<ErrorValue>(evaluatorResult);
-            }
-        }
-
-        private void TestBindingError(string script, Features features, string errorMessage, SymbolTable symbolTable = null)
-        {
-            var config = new PowerFxConfig(features)
-            {
-                SymbolTable = symbolTable
-            };
-
-            var engine = new Engine(config);
-            var result = engine.Check(script);
-
-            Assert.False(result.IsSuccess);
-
-            Assert.Contains(result.Errors, error => error.Message.Contains(errorMessage));
-        }
     }
 }
